+++ core/enums/RefreshMode.js
@@ -6,8 +6,10 @@
  */
 
 /**
- * Options for how tabs should be refreshed via their built-in RefreshContextModel when they receive
+ * Options for how content should be refreshed via their built-in RefreshContextModel when they receive
  * a refresh request while inactive/hidden or are (re)activated by their parent container.
+ *
+ * Used by TabContainerModel and DashContainerModel.
  *
  * @enum {RefreshMode}
  */