+++ core/enums/RenderMode.js
@@ -6,9 +6,10 @@
  */
 
 /**
- * Options for how tab contents should be rendered by their parent container. A key feature of
- * TabContainer and DashContainer is that they will by default render child tabs lazily,
- * only when activated for the first time (LAZY)
+ * Options for how contents should be rendered by their parent container.
+ *
+ *
+ * Used by TabContainerModel, DashContainerModel, and PanelModel.
  *
  * @enum {RenderMode}
  */
@@ -17,7 +18,7 @@ export const RenderMode = Object.freeze({
     /** Always render contents when the parent container is rendered, even if inactive. */
     ALWAYS: 'always',
 
-    /** Render contents lazily or "just in time" - only if/when the tab is first activated. */
+    /** Render contents lazily or "just in time" - only if/when the content is first activated. */
     LAZY: 'lazy',
 
     /** Render lazily, and actively unmount the contents if/when de-activated. */