+++ desktop/cmp/tab/impl/Tab.js
@@ -5,10 +5,9 @@
  * Copyright © 2020 Extremely Heavy Industries Inc.
  */
 import {useRef} from 'react';
-import {elem, hoistCmp, uses, ModelPublishMode} from '@xh/hoist/core';
+import {elem, hoistCmp, uses, ModelPublishMode, RenderMode} from '@xh/hoist/core';
 import {refreshContextView} from '@xh/hoist/core/refresh';
 import {frame} from '@xh/hoist/cmp/layout';
-import {RenderMode} from '@xh/hoist/enums';
 import {TabModel} from '@xh/hoist/cmp/tab';
 
 /**