+++ mobile/cmp/tab/impl/Tab.js
@@ -5,10 +5,9 @@
  * Copyright © 2020 Extremely Heavy Industries Inc.
  */
 import {useRef} from 'react';
-import {elem, hoistCmp, ModelPublishMode, uses} from '@xh/hoist/core';
+import {elem, hoistCmp, ModelPublishMode, RenderMode, uses} from '@xh/hoist/core';
 import {refreshContextView} from '@xh/hoist/core/refresh';
 import {page as onsenPage} from '@xh/hoist/kit/onsen';
-import {RenderMode} from '@xh/hoist/enums';
 
 import {TabModel} from '@xh/hoist/cmp/tab';